Switch to Use MDX Rather than Remark (#4)
Remark doesn't allow you to use React components within Markdown posts and we want to use this functionality. Therefore, following the Gatsby upgrade guide, we have migrated over to use MDX that gives us this functionality.

We have also started to use the frontmatter `Description` property for the blog listing excepts rather than the start of the post itself.